Split OAuth test into two files
